requirements.txt
certifi==2025.11.12
charset-normalizer==3.4.4
idna==3.11
markdown-it-py==4.0.0
mdurl==0.1.2
packaging==25.0
Pygments==2.19.2
pynetbox==7.5.0
python-dotenv==1.2.1
requests==2.32.5
rich==14.2.0
urllib3==2.6.2
